Specify publisher and provider addresses (#23)
* Specify publisher and provider addresses

The publisher address is used to tell IPNI where to fetch advertisements from. It may be a public address different from the publisher listen address. So, it is necessary to explicitly specify the address.

The provider address is the provider address part of an advertisement. IPNI may require that this be a routable address and reject the advertisement if it is not. So, it is necessary to specify this address.

* mod tidy